Centralized express server & Various improvements
