+++ .env-sample
@@ -1,11 +1,16 @@
 NODE_ENV=production
 
-SUPINBOT_REDIS_URI=redis://redis:6379/1
+SUPINBOT_REDIS_URI=redis://redis:6379
 
 SUPINBOT_SLACK_TOKEN=xoxb-00000000000-XXXXXXXXXXXXXXXXXXXXXXXX
 SUPINBOT_SLACK_NAME=SUPINBOT
 SUPINBOT_SLACK_PARAMM={"icon_emoji": ":robot_face:"}
 
 SUPINBOT_LOG_LEVEL=info
+
+SUPINBOT_WEB_URL=http://localhost:8080/
+SUPINBOT_WEB_TRUST_PROXY=""loopback""
+SUPINBOT_WEB_PORT=8080
+SUPINBOT_WEB_DISABLE_CACHE=false
 
 SUPINBOT_PLUGINS=[]
