+++ .jshintrc
@@ -1,0 +1,5 @@
+{
+	"strict": "global",
+	"esversion": 6,
+	"node": true
+}
