+++ lib/config.js
@@ -1,4 +1,6 @@
-var convict = require('convict');
+'use strict';
+
+const convict = require('convict');
 
 var config = convict({
 	env: {
@@ -18,28 +20,52 @@ var config = convict({
 			doc: 'Name of the BOT',
 			format: String,
 			default: 'SUPINBOT',
-                        env: 'SUPINBOT_SLACK_NAME'
+			env: 'SUPINBOT_SLACK_NAME'
 		},
 		params: {
 			doc: 'Default parameters to send along with messages',
 			format: Object,
 			default: '{"icon_emoji": ":robot_face:"}',
-                        env: 'SUPINBOT_SLACK_PARAMS'
+			env: 'SUPINBOT_SLACK_PARAMS'
 		}
 	},
-	log: {
-		level: {
-			doc: 'The winston logging level',
-			format: ['silly', 'debug', 'verbose', 'info', 'warn', 'error'],
-			default: 'info',
-			env: 'SUPINBOT_LOG_LEVEL'
-		}
+	log_level: {
+		doc: 'The winston logging level',
+		format: ['silly', 'debug', 'verbose', 'info', 'warn', 'error'],
+		default: 'info',
+		env: 'SUPINBOT_LOG_LEVEL'
 	},
 	redis: {
 		doc: 'The Redis connection URI',
 		format: String,
-		default: 'redis://redis:6379/1',
+		default: 'redis://redis:6379',
 		env: 'SUPINBOT_REDIS_URI'
+	},
+	web: {
+		url: {
+			doc: 'The URL to the web interface (with trailing /)',
+			format: String,
+			default: 'http://localhost:8080/',
+			env: 'SUPINBOT_WEB_URL'
+		},
+		trust_proxy: {
+			doc: 'See: https://expressjs.com/en/guide/behind-proxies.html',
+			format: Object,
+			default: '"loopback"',
+			env: 'SUPINBOT_WEB_TRUST_PROXY'
+		},
+		port: {
+			doc: 'The port the web server will listen on',
+			format: 'nat',
+			default: 8080,
+			env: 'SUPINBOT_WEB_PORT'
+		},
+		no_cache: {
+			doc: 'Disable the caching of templates',
+			format: Boolean,
+			default: false,
+			env: 'SUPINBOT_WEB_DISABLE_CACHE'
+		}
 	},
 	plugins: {
 		doc: 'A JSON Array of enabled plugin names',
