+++ package.json
@@ -15,7 +15,9 @@
     "@slack/client": "^3.1.0",
     "convict": "^1.0.2",
     "dotenv": "^2.0.0",
+    "express": "^4.14.0",
     "extend": "^3.0.0",
+    "nunjucks": "^2.4.2",
     "winston": "^2.1.1"
   }
 }
